+++ src/main/java/com/trident/security/service/InVulService.java
@@ -17,13 +17,13 @@ public interface InVulService {
     List<InVul> getAllInVul(Integer pageNum, Integer pageSize) throws Exception;
     List<InVul> getAllActiveInVul(Integer pageNum, Integer pageSize) throws Exception;
     List<InVul> getAllInVulByStatus(Integer status,Integer pageNum, Integer pageSize) throws Exception;
-    InVul getInVulById(Integer id) throws Exception;
+    InVul getInVulById(Long id) throws Exception;
     List<InVul> getInVulByExample(InVul inVul, Date createEndTime, Date expectEndTime, Integer pageNum, Integer pageSize) throws Exception;
     
     Integer addInVul(InVul invul) throws Exception;
     Integer updateInVul(InVul inVul) throws Exception;
-    Integer deleteInVulById(Integer id) throws Exception;
-    Integer changeStatusById(Integer id,Integer status) throws Exception;
+    Integer deleteInVulById(Long id) throws Exception;
+    Integer changeStatusById(Long id,Integer status) throws Exception;
     
     Integer getAllCount() throws Exception;
     Integer getAvailableCountByStatus(Integer status) throws Exception;
